+++ integrations/github/main.py
@@ -14,10 +14,9 @@ from github.webhook.webhook_processors.repository_webhook_processor import (
 )
 from github.webhook.webhook_client import GithubWebhookClient
 from github.core.exporters.repository_exporter import RestRepositoryExporter
-from github.core.options import ListRepositoryOptions
+from github.core.options import ListPullRequestOptions, ListRepositoryOptions
 from typing import TYPE_CHECKING
 from port_ocean.context.event import event
-from port_ocean.utils.async_iterators import stream_async_iterators_tasks
 from github.webhook.webhook_processors.issue_webhook_processor import (
     IssueWebhookProcessor,
 )
@@ -26,6 +25,15 @@ from github.core.options import ListIssueOptions
 
 if TYPE_CHECKING:
     from integration import GithubPortAppConfig, GithubIssueConfig
+from github.core.exporters.pull_request_exporter import RestPullRequestExporter
+from port_ocean.utils.async_iterators import stream_async_iterators_tasks
+from github.webhook.webhook_processors.pull_request_webhook_processor import (
+    PullRequestWebhookProcessor,
+)
+
+
+if TYPE_CHECKING:
+    from integration import GithubPortAppConfig, GithubPullRequestConfig
 
 
 @ocean.on_start()
@@ -73,6 +81,36 @@ async def resync_repositories(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         yield repositories
 
 
+@ocean.on_resync(ObjectKind.PULL_REQUEST)
+async def resync_pull_requests(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    """Resync all pull requests in the organization's repositories."""
+    logger.info(f"Starting resync for kind: {kind}")
+
+    rest_client = create_github_client()
+    repository_exporter = RestRepositoryExporter(rest_client)
+    pull_request_exporter = RestPullRequestExporter(rest_client)
+    config = cast("GithubPullRequestConfig", event.resource_config)
+
+    repo_options = ListRepositoryOptions(
+        type=cast("GithubPortAppConfig", event.port_app_config).repository_type
+    )
+
+    async for repos in repository_exporter.get_paginated_resources(
+        options=repo_options
+    ):
+        tasks = [
+            pull_request_exporter.get_paginated_resources(
+                ListPullRequestOptions(
+                    repo_name=repo["name"],
+                    state=config.selector.state,
+                )
+            )
+            for repo in repos
+        ]
+        async for pull_requests in stream_async_iterators_tasks(*tasks):
+            yield pull_requests
+
+
 @ocean.on_resync(ObjectKind.ISSUE)
 async def resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     """Resync all issues from repositories."""
@@ -104,4 +142,5 @@ async def resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
 
 
 ocean.add_webhook_processor("/webhook", RepositoryWebhookProcessor)
+ocean.add_webhook_processor("/webhook", PullRequestWebhookProcessor)
 ocean.add_webhook_processor("/webhook", IssueWebhookProcessor)